Added exec delay and wall commands to updateos script

+++ updos.sh
@@ -1,11 +1,15 @@
 ########################################################
 # [AUTOMATED-INSERT-MARKER]
 # Author Eric Johnfelt
-# Date 10/23/2017
+# Date 7/30/2019
 
 # Update OS
 function UpdateOS()
 {
+	CMD=""
+	NOTIFY="no"
+	REPLYDELAY="60"
+	EXECDELAY=""
 	PREFIX=""
 
 	if [ ! "${LOGNAME}" = "root" ]; then
@@ -14,7 +18,31 @@ function UpdateOS()
 
 	GetPackageManager
 
-	if [ ! "$1" = "-c" ]; then
+	while [ ! "${1}" = "" ]; do
+		case "${1}" in
+		"-c") CMD=check ;;
+		"-w") CMD=wait ;;
+		"-r") CMD=reboot ;;
+		"-h") CMD=halt ;;
+		"-n") NOTIFY="yes" ;;
+		"-t") REPLYDELAY="${2}"; shift 1 ;;
+		"-e") EXECDELAY="${2}"; shift 1 ;;
+		esac
+
+		shift 1
+	done
+
+	if [ ! "${EXECDELAY}" = "" ]; then
+		sleep ${EXECDELAY}
+	fi
+
+	if [ "${NOTIFY}" = "yes" ]; then
+		${PREFIX} wall "The system is about to patch, it will likely reboot, you have 5 minutes to finish what you are doing"
+
+		sleep 5m
+	fi
+
+	if [ ! "${CMD}" = "check" ]; then
 		echo -e "Beginning Update of ${HOSTNAME}..."
 		cmds=${#UPDCMDS[*]}
 		index=0
@@ -23,25 +51,23 @@ function UpdateOS()
 			eval ${PREFIX} ${UPDCMDS[${index}]}
 			index=$((${index} + 1))
 		done
-	fi
 
-	if [ "$1" = "-c" ]; then
-		chkupd
-	elif [ "$1" = "-w" ]; then
-		read -p "Reboot ${HOSTNAME} (y/n)? "
-		[ "${REPLY}" = "y" ] && ${PREFIX} reboot
-	elif [ "$1" = "-r" -o "$1" = "-y" ]; then
-		read -n 1 -t 30 -p "Rebooting ${HOSTNAME} in 30s, abort (y/n)? "
-		[ ! "${REPLY}" = "y" ] && ${PREFIX} reboot
-	elif [ "$1" = "-h" -o "$1" = "-s" ]; then
-		read -n 1 -t 30 -p "Shutting down ${HOSTNAME} in 30s, abort (y/n)? "
-		[ ! "${REPLY}" = "y" ] && ${PREFIX} shutdown -h now
+		case "${CMD}" in
+		"wait")
+			read -p "Reboot ${HOSTNAME} (y/n)? "
+			[ "${REPLY}" = "y" -o "${REPLY}" = "Y" ] && ${PREFIX} reboot ;;
+		"reboot")
+			read -n 1 -t ${REPLYDELAY} -p "Rebooting ${HOSTNAME} in ${REPLYDELAY}s, abort (y/n)? "
+			[ ! "${REPLY}" = "y" ] && ${PREFIX} reboot ;;
+		"halt")
+			read -n 1 -t ${REPLYDELAY} -p "Shutting down ${HOSTNAME} in ${REPLYDELAY}s, abort (y/n)? "
+			[ ! "${REPLY}" = "y" ] && ${PREFIX} shutdown -h now ;;
+		*)
+			read -t 60 -p "Reboot ${HOSTNAME} (y/N)? "
+			[ "${REPLY}" = "y" -o "${REPLY}" = "Y" ] && ${PREFIX} reboot
+		esac
 	else
-		read -t 60 -p "Reboot ${HOSTNAME} (y/N)? "
-
-		if [ "${REPLY}" = "y" -o "${REPLY}" = "Y" ]; then
-			${PREFIX} reboot
-		fi
+		chkupd
 	fi
 }
